+++ packages/sequence-diagram/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@appland/sequence-diagram-v1.8.0](https://github.com/getappmap/appmap-js/compare/@appland/sequence-diagram-v1.7.0...@appland/sequence-diagram-v1.8.0) (2023-09-07)
+
+
+### Features
+
+* Sequence diagram validation ([e88ab44](https://github.com/getappmap/appmap-js/commit/e88ab44791a994f4102903e8141e538db481d23a))
+
 # [@appland/sequence-diagram-v1.7.0](https://github.com/getappmap/appmap-js/compare/@appland/sequence-diagram-v1.6.1...@appland/sequence-diagram-v1.7.0) (2023-07-19)